refactor(n8n-cli): Address review comments on naming and doc clarity
